pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.12.4",
    "discord-py>=2.5.2",
    "google-generativeai>=0.8.5",
    "python-dotenv>=1.1.0",
]
